added validate_certs=False to nb_lookup
This wasn't needed while I was testing because I had an environment
variable for "export REQUESTS_CA_BUNDLE=/path/to/ca.crt" which was
causing pynetbox to properly validate certificates.

For the purposes of this example I've set everything to not validate,
but in production you may want to do this properly.